pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iris"
version = "1.0.0"
description = "Machine learning pipeline for cell cycle prediction in IRIS cell imaging data"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "scikit-image>=0.25.2",
    "matplotlib>=3.10.6",
    "h5py>=3.14.0",
    "scipy>=1.15.2",
    "tqdm>=4.67.1",
    "lightning[pytorch-extra]>=2.5.5",
    "mlflow>=2.8.0",
    "pyyaml>=6.0",
    "tensorboard>=2.14.0",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "seaborn>=0.13.2",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/andfrey/iris"
Repository = "https://github.com/andfrey/iris"
Documentation = "https://github.com/andfrey/iris/tree/master/docs"
Issues = "https://github.com/andfrey/iris/issues"

[tool.setuptools]
packages = [".src", ".src.data_pipeline", ".src.train", ".src.models"]

[tool.setuptools.package-data]
src = ["py.typed"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "fast: marks tests as fast (select with '-m fast')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "seaborn (>=0.13.2,<0.14.0)"
]
